separate unsupported generics
